requirements.txt
appdirs==1.4.3
click==6.7
cycler==0.10.0
decorator==4.0.11
Flask==0.12.1
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
matplotlib==2.0.0
numpy==1.12.1
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.1
